% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxypackagedoc-package.R
\docType{package}
\name{roxypackagedoc-package}
\alias{roxypackagedoc-package}
\alias{_PACKAGE}
\title{Testing erroneous alias generation}
\description{
Test for roxygen2 error when generating pkgname-package man page.
}
\keyword{internal}
